Fixed blog and category template page bugs
