+++ src/components/footer/footer.js
@@ -1,6 +1,6 @@
 import './footer.css';
 
-import ContactForm2 from '../form2';
+import ContactForm from '../form';
 import {Link} from 'gatsby';
 import React from 'react';
 import facebook from '../../images/iconography/facebook.png';
@@ -14,7 +14,7 @@ const Footer = () => {
         <hr/>
         <div className='content-wrapper'>
             <div className='left'>
-                <ContactForm2 />
+                <ContactForm />
             </div>
             <div className='right'>
                 <div className='helpful-links'>
