+++ src/templates/tagsTemplate.js
@@ -77,7 +77,7 @@ const Tags = ({ pageContext, data }) => {
       return setPosts(newPostList);
     }
     renderInitialPosts();
-  });
+  }, []);
 
   //write a function that will update state to show 6 more posts
   const showMorePosts = (clickCount, isFinished) => {
